\name{RevoHStream-package}
\alias{RevoHStream-package}
\docType{package}
\title{rhstream: Streaming Wrappers for R and Hadoop}
\description{
	Provides simple wrappers to compute \code{tapply}and \code{lapply} using Hadoop, along
with basic map/reduce functionality.
}
\seealso{
rhstream,rhtapply,rhlapply
}
