\name{rhbase}
\alias{rhbase}
\alias{rhbase}
\docType{package}
\title{rhbase: Interact with HBase tables from R}
\description{
   The functions allows one to create or delete tables and insert, delete, and scan records
   in HBase tables.  
}
\seealso{
hb.insert,
hb.get,
hb.delete,
hb.scan,
hb.insert.data.frame,
hb.get.data.frame,
hb.defaults,
hb.init,
hb.defaults,
hb.init,
hb.list.tables,
hb.describe.table,
hb.set.table.mode,
hb.delete.table,
hb.new.table,
hb.regions.table
}

