\name{dfs}
\alias{dfs.ls}
\alias{dfs.get}
\alias{dfs.put}
\alias{dfs.rm}
\alias{dfs.rmr}
\alias{dfs.exists}

\title{Utility functions for working with HDFS}
\description{Utility functions for working with HDFS
}

\usage{
	dfs = function(cmd, ...)
	dfs.ls = dfs.match
	dfs.get = dfs.match
	dfs.put = dfs.match
	dfs.rm = dfs.match
	dfs.rmr = dfs.match
	dfs.exists = function(f)
}

\arguments{
  \item{cmd}{HDFS command to execute}
  \item{f}{path to HDFS file}
}

\details{ }
  
\examples{  }