\name{keyval}
\alias{getKeys}
\alias{getValues}

\title{Utility function}
\description{Utility function
}
	
\usage{
	getKeys = function(l)
	getValues = function(l)
	keyval = function(k, v = NULL, i = 1)
}

\arguments{
  \item{l}{list containing key/value pairs}
  \item{k}{the desired key or a list of lent}
  \item{v}{}
  \item{i}{}
}

\details{ }

\examples{  }