\name{rawtextinputformat}

\title{Utility function that reads a raw line of text from stdin}
\description{Utility function that reads a raw line of text from stdin}

\usage{
rawtextinputformat(line) 
}

\arguments{
  \item{line}{The line of text to read from stdin}
}

\details{ }

\examples{  }