\name{RevoHStream}
\alias{RevoHStream}
\docType{package}
\title{A package to perform Map Reduce computations in R}
\description{Running on top of Hadoop, this package allows to define an run map reduce jobs, including specifying the mapper and the reducer as R functions and to move data between R and hadoop in a mostly transparent way. The aim is to make writing map reduce jobs very similar and just as easy as writing a lapply and a tapply. Additional features provide easy job composition, transparent intermediate result management, support for mapper and reducer definition and more.
}
